Good stuff

+++ src/kernel/kwin/frame.cpp
@@ -6,7 +6,7 @@ kwin::FrameTest::draw( kwin::Context &ctx )
 {
     ctx.rectOutline(m_tl, m_sp, m_col);
 
-    ctx.rect(m_tl+ivec2(24), m_sp-ivec2(48), m_col);
+    ctx.rect(m_tl+vec2(24), m_sp-vec2(48), m_col);
 }
